// base/free5gc/webconsole/frontend/src/api/api.ts
export interface ANInformation {
    /**
     * 
     * @type {string}
     * @memberof ANInformation
     */
    'IPAddress': string;
    /**
     * 
     * @type {number}
     * @memberof ANInformation
     */
    'TEID': number;
}
// ...
export interface Dnn {
    /**
     * 
     * @type {string}
     * @memberof Dnn
     */
    'dnn': string;
}
// ...
export interface PduSession {
    /**
     * 
     * @type {string}
     * @memberof PduSession
     */
    'Dnn': string;
    /**
     * 
     * @type {string}
     * @memberof PduSession
     */
    'PduSessionId': string;
    /**
     * 
     * @type {string}
     * @memberof PduSession
     */
    'Sd': string;
    /**
     * 
     * @type {string}
     * @memberof PduSession
     */
    'SmContextRef': string;
    /**
     * 
     * @type {string}
     * @memberof PduSession
     */
    'Sst': string;
}
/**
 * 
 * @export
 * @interface PduSessionInfo
 */
export interface PduSessionInfo {
    /**
     * 
     * @type {string}
     * @memberof PduSessionInfo
     */
    'AnType': string;
    /**
     * 
     * @type {string}
     * @memberof PduSessionInfo
     */
    'Dnn': string;
    /**
     * 
     * @type {string}
     * @memberof PduSessionInfo
     */
    'LocalSEID'?: string;
    /**
     * 
     * @type {string}
     * @memberof PduSessionInfo
     */
    'PDUAddress': string;
    /**
     * 
     * @type {string}
     * @memberof PduSessionInfo
     */
    'PDUSessionID': string;
    /**
     * 
     * @type {string}
     * @memberof PduSessionInfo
     */
    'RemoteSEID'?: string;
    /**
     * 
     * @type {string}
     * @memberof PduSessionInfo
     */
    'Sd': string;
    /**
     * 
     * @type {SessionRule}
     * @memberof PduSessionInfo
     */
    'SessionRule': SessionRule;
    /**
     * 
     * @type {string}
     * @memberof PduSessionInfo
     */
    'Sst': string;
    /**
     * 
     * @type {string}
     * @memberof PduSessionInfo
     */
    'Supi': string;
    /**
     * 
     * @type {Tunnel}
     * @memberof PduSessionInfo
     */
    'Tunnel': Tunnel;
    /**
     * 
     * @type {string}
     * @memberof PduSessionInfo
     */
    'UpCnxState': string;
}
// ...
export interface SessionRule {
    /**
     * 
     * @type {string}
     * @memberof SessionRule
     */
    'sessRuleId': string;
}
// ...
export interface Tunnel {
    /**
     * 
     * @type {ANInformation}
     * @memberof Tunnel
     */
    'ANInformation': ANInformation;
    /**
     * github.com/free5gc/smf/internal/context/datapath.go:DataPathPool
     * @type {any}
     * @memberof Tunnel
     */
    'DataPathPool': any;
    /**
     * github.com/free5gc/util/idgenerator/idgenerator.go:IDGenerator
     * @type {any}
     * @memberof Tunnel
     */
    'PathIDGenerator': any;
}
/**
 * 
 * @export
 * @interface UeContext
 */
export interface UeContext {
    /**
     * 
     * @type {string}
     * @memberof UeContext
     */
    'AccessType': string;
    /**
     * 
     * @type {string}
     * @memberof UeContext
     */
    'CmState': string;
    /**
     * 
     * @type {string}
     * @memberof UeContext
     */
    'Guti': string;
    /**
     * 
     * @type {string}
     * @memberof UeContext
     */
    'Mcc': string;
    /**
     * 
     * @type {string}
     * @memberof UeContext
     */
    'Mnc': string;
    /**
     * 
     * @type {Array<PduSession>}
     * @memberof UeContext
     */
    'PduSessions': Array<PduSession>;
    /**
     * 
     * @type {string}
     * @memberof UeContext
     */
    'Supi': string;
    /**
     * 
     * @type {string}
     * @memberof UeContext
     */
    'Tac': string;
}